# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flares",
    "description": "HTTP issue tracking with Pushover notifications",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/alerts": {
      "post": {
        "tags": [],
        "summary": "Send a one-shot notification without creating or updating an issue.",
        "operationId": "send_alert",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Retry key; reuse with the same alert body",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Alert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResult"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
    "/v1/heartbeat": {
      "delete": {
        "tags": [],
        "operationId": "delete_heartbeat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Monitor deleted"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/heartbeats": {
      "get": {
        "tags": [],
        "operationId": "list_heartbeats",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Heartbeat"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      },
      "post": {
        "tags": [],
        "operationId": "register_heartbeat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Heartbeat"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/heartbeats/check-in": {
      "post": {
        "tags": [],
        "operationId": "check_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseIssue"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Heartbeat"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/issue": {
      "get": {
        "tags": [],
        "operationId": "lookup_issue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Issue"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/issues": {
      "get": {
        "tags": [],
        "operationId": "list_issues",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/IssueStatus"
                }
              ]
            }
          },
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100,
              "maximum": 1000,
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueList"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/issues/close": {
      "post": {
        "tags": [],
        "operationId": "close_issue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseIssue"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/issues/open": {
      "post": {
        "tags": [],
        "operationId": "open_issue",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenIssue"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/v1/issues/{id}": {
      "get": {
        "tags": [],
        "operationId": "get_issue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "URL-encoded issue ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Issue"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Alert": {
        "type": "object",
        "required": [
          "title",
          "message"
        ],
        "properties": {
          "group_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "title": {
            "type": "string",
            "maxLength": 250,
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "AlertResult": {
        "type": "object",
        "required": [
          "delivery_id",
          "notification"
        ],
        "properties": {
          "delivery_id": {
            "type": "integer",
            "format": "int64"
          },
          "notification": {
            "$ref": "#/components/schemas/Notification"
          }
        }
      },
      "CloseIssue": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
      "Heartbeat": {
        "allOf": [
          {
            "$ref": "#/components/schemas/HeartbeatInput"
          },
          {
            "type": "object",
            "required": [
              "last_seen",
              "due_at",
              "overdue"
            ],
            "properties": {
              "due_at": {
                "type": "integer",
                "format": "int64"
              },
              "last_seen": {
                "type": "integer",
                "format": "int64"
              },
              "overdue": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "HeartbeatInput": {
        "type": "object",
        "required": [
          "id",
          "title",
          "interval_seconds"
        ],
        "properties": {
          "grace_seconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "interval_seconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "notify_on_recovery": {
            "type": "boolean"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "title": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Issue": {
        "type": "object",
        "required": [
          "severity",
          "notify_on_resolution",
          "id",
          "status",
          "title",
          "message",
          "created_at",
          "updated_at",
          "opened_at",
          "opening_count",
          "notification"
        ],
        "properties": {
          "closed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "delivery_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "notification": {
            "$ref": "#/components/schemas/Notification"
          },
          "notify_on_resolution": {
            "type": "boolean"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time"
          },
          "opening_count": {
            "type": "integer",
            "format": "int64"
          },
          "remind_every_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "IssueList": {
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Issue"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "IssueStatus": {
        "type": "string",
        "enum": [
          "open",
          "closed"
        ]
      },
      "MutationResult": {
        "type": "object",
        "required": [
          "issue",
          "changed",
          "notification"
        ],
        "properties": {
          "changed": {
            "type": "boolean"
          },
          "delivery_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "issue": {
            "$ref": "#/components/schemas/Issue"
          },
          "notification": {
            "$ref": "#/components/schemas/Notification",
            "description": "The attempt made by this request, not the latest historical outcome."
          }
        }
      },
      "Notification": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/NotificationStatus"
          }
        }
      },
      "NotificationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "sent",
          "failed",
          "unknown",
          "not_attempted"
        ]
      },
      "OpenIssue": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 1024,
            "minLength": 1
          },
          "notify_on_resolution": {
            "type": "boolean"
          },
          "remind_every_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 250,
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "Severity": {
        "type": "string",
        "enum": [
          "info",
          "warning",
          "critical"
        ]
      }
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}